feat: 결제 상태 관리 기능 추가 및 예약 연동 개선
- PaymentStatus 열거형 추가로 결제 상태(PENDING, SUCCESS, AWAIT, FAILED) 관리
- Payment 엔티티 수정 및 관련 정적 팩토리 메소드 추가(pending, await)
- 결제 성공/실패 처리 로직 도입(success, fail 메소드)
- BigDecimal 타입으로 결제 금액 데이터 형식 변경
- 예약 승인 시 대기 상태 결제 정보 생성 및 저장 로직(paymentService.await) 추가
- TossPaymentException이 상위 PaymentException 상속하도록 리팩토링
- JS 결제 데이터 생성 및 API 요청에 대한 로직 개선

+++ src/main/java/roomescape/payment/application/dto/PaymentRequest.java
@@ -1,7 +1,9 @@
 package roomescape.payment.application.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentRequest(
-        Long amount,
+        BigDecimal amount,
         String orderId,
         String paymentKey
 ) {
